requirements.txt
# VS Code SageMaker Development Requirements
# Core AWS & SageMaker
boto3>=1.34.0
sagemaker>=2.189.0
# awscli handled via conda (pip only ships v1)

# Environment management
python-dotenv>=1.0.0

# ML Libraries (for local development)
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
datasets>=2.14.0
accelerate>=0.24.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
wordfreq>=3.1.1
sentence-transformers>=2.7.0
uvicorn>=0.32.0
fastapi>=0.115.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.0.0
pytest>=7.4.0

# Optional: Local model testing
llama-cpp-python>=0.2.0
bitsandbytes>=0.41.0

gradio
mcp-tools>=0.1.0

# Voice Pipeline - STT
faster-whisper>=0.10.0
ctranslate2>=4.0.0

# Voice Pipeline - TTS
TTS>=0.22.0                        # Coqui XTTS v2

# Voice Pipeline - Wake Word
openwakeword>=0.5.0
onnxruntime>=1.16.0

# Voice Pipeline - Audio I/O
sounddevice>=0.4.6
soundfile>=0.12.1
webrtcvad>=2.0.10
PyYAML>=6.0.0
